pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "energyallocation"
version = "0.1.0"
description = "Energy allocation platform foundation: API contract, alerting model, and ROI calculator"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
	"fastapi>=0.116.0",
	"httpx>=0.28.0",
	"psycopg[binary]>=3.2.0",
	"PyJWT>=2.9.0",
	"uvicorn[standard]>=0.35.0",
]

[project.scripts]
energy-roi = "roi_calculator.cli:main"
energy-api = "energy_api.main:run"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
